Worked out a few bugs found while testing with Tamra, including lost cards on disconnect by storing client hands server-side and garbage collection after timeout based on a three-second heartbeat interval. Max missed heartbeats is 5, for a total of 15 seconds before declaring connection lost. This was used because android sockets do not release blocked system calls when a socket dies (as opposed to being closed).
Also made HandActivity more dynamic by creating buttons on the fly instead of using five pre-made. However to do this I had to generate ids for each button, and I worry about collisions from the random number generator.

Finally, added a quit button to the player screen which discards the hand and closes the socket, for  a clean disconnect.